+++ changelog.md
@@ -1,5 +1,5 @@
 ## 6.1.3.0
-* Added message to chat when reloading something fails
+* Added message to chat when reloading generates errors or warnings
 
 ## 6.1.2.0
 * Inverted how `priority` on mobs works.